Restart just the audio service instead of the entire computer.
There is no need to restart the entire machine when the device
configuration is changed - the only thing that needs restarting is the
Windows Audio service.

This code is quite naive and doesn't really handle errors nor dependent
services well, but it's a start.

This resolves https://sourceforge.net/p/equalizerapo/tickets/23/